Separating views.py to files, kubernetes configuration
Moved functions from views.py to project_views.py, issue_views.py, subissue_views.py and label_views.py. Added volume and volume claim yaml files used for media files. Renamed some kubernetes configuration files.
