+++ kubernetes_config/postgres-deployment.yaml
@@ -42,3 +42,15 @@ spec:
         - name: postgres-volume-mount
           persistentVolumeClaim:
             claimName: postgres-pvc
+---
+kind: Service
+apiVersion: v1
+metadata:
+  name: postgres
+spec:
+  selector:
+    app: postgres-container
+  ports:
+    - protocol: TCP
+      port: 5432
+      targetPort: 5432
