--- kubernetes_config/postgres-service.yaml
@@ -1,11 +1,0 @@
-kind: Service
-apiVersion: v1
-metadata:
-  name: postgres
-spec:
-  selector:
-    app: postgres-container
-  ports:
-    - protocol: TCP
-      port: 5432
-      targetPort: 5432
