+++ uks_app/views/__init__.py
@@ -1,0 +1,5 @@
+from .project_views import *
+from .issue_views import *
+from .subissue_views import *
+from .label_views import *
+from .other_views import *
