+++ uks_app/views/other_views.py
@@ -4,14 +4,14 @@ from django.views import generic
 from django.urls import reverse_lazy
 from django.views.generic.edit import DeleteView
 
-from .models import ObservedProject, Issue, Milestone, MilestoneChange, Event, Label, User, Comment, CommentChange, CodeChangeEvent, AssignIssueEvent, IssueChange, LableEvent, SubIssueEvent
-from .forms import ProjectForm, IssueForm, MilestoneForm, LabelForm, ChooseLabelForm, UserRegisterForm, UserUpdateForm, ProfileUpdateForm, ChooseMilestoneForm, CommentForm, ChooseSubissueForm, AssignIssueForm
+from uks_app.models import ObservedProject, Issue, Milestone, MilestoneChange, Event, Label, User, Comment, CommentChange, CodeChangeEvent, AssignIssueEvent, IssueChange, LableEvent, SubIssueEvent
+from uks_app.forms import ProjectForm, IssueForm, MilestoneForm, LabelForm, ChooseLabelForm, UserRegisterForm, UserUpdateForm, ProfileUpdateForm, ChooseMilestoneForm, CommentForm, ChooseSubissueForm, AssignIssueForm
 import logging
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
-from .models import ObservedProject, Issue, CodeChange
-from .forms import ProjectForm, IssueForm, UserRegisterForm, UserUpdateForm, ProfileUpdateForm
+from uks_app.models import ObservedProject, Issue, CodeChange
+from uks_app.forms import ProjectForm, IssueForm, UserRegisterForm, UserUpdateForm, ProfileUpdateForm
 from django.views.decorators.http import require_http_methods
 from django.views.decorators.csrf import csrf_exempt
 import json
@@ -30,325 +30,9 @@ from django.views.decorators.cache import cache_page
 
 CACHE_TTL = getattr(settings, 'CACHE_TTL', DEFAULT_TIMEOUT)
 
-
 # home page
 def index(request): 
     return render(request, 'uks_app/index.html') #optional third argument context
-
-# new project form
-@login_required
-def create_update_project(request, project_id=None):
- 
-    user = request.user
-
-    #project_id == None -> Create
-    #project_id != None -> Update
-    observed_project = get_object_or_404(ObservedProject, pk=project_id) if project_id else None
-
-    # only project owner can make changes to the project
-    if project_id != None and observed_project.user != request.user:
-        return HttpResponse('Unauthorized', status=401)
-
-    form = ProjectForm(request.POST or None, instance=observed_project)
-
-    if form.is_valid():
-         
-        project = form.save(commit=False) 
- 
-        project.user = user
-        project.save()
-
-        if project_id:
-            return HttpResponseRedirect('/project/' + str(project_id) + '/')  
-        else:
-            return HttpResponseRedirect('/project/' + str(project.id) + '/') 
-
-    context = {
-        'form' : form,
-    }
-
-    return render(request, 'uks_app/create_update_project.html', context)
-
-# new issue form
-@login_required
-def create_update_issue(request, project_id, issue_id=None):
-
-    #get observed project
-    observed_project = get_object_or_404(ObservedProject, id=project_id)
-
-    #issue_id == None -> Create
-    #issue_id != None -> Update
-    observed_issue = get_object_or_404(Issue, pk=issue_id) if issue_id else None
-
-    # only project owner and collaborators can create a public issue
-    if issue_id == None and observed_project.public == False and request.user != observed_project.user and not observed_project.collaborators.filter(id = request.user.id).exists():
-        return HttpResponse('Unauthorized', status=401)
-    
-    # only project owner and collaborators can update a issue
-    if issue_id != None and request.user != observed_project.user and not observed_project.collaborators.filter(id = request.user.id).exists():
-        return HttpResponse('Unauthorized', status=401)
-
-    form = IssueForm(data=request.POST or None, project=observed_project, iss=observed_issue, instance=observed_issue)
-
-    if form.is_valid():
-        issue = form.save(commit=False)
-       
-        if not issue_id:
-            issue.create_time = datetime.datetime.now()
-        
-        #set project as foreign key
-        issue.project = observed_project
-        issue.save()
-        
-        if issue_id == None and issue.parent_issue != None :
-            SubIssueEvent.objects.create(user=request.user, time= datetime.datetime.now(), issue=issue.parent_issue, state="CR", subissue=issue)
-
-
-        if issue_id:
-            return HttpResponseRedirect('/issue/' + str(issue_id) + '/')  
-        else:
-            return HttpResponseRedirect('/issue/' + str(issue.id) + '/')  
-
-    cancel_url = "../../../../../issue/"+str(issue_id) if issue_id else "../../"
-
-    
-    context = {
-        'form' : form,
-        'cancel_url': cancel_url
-    }
-
-    return render(request, 'uks_app/create_update_issue.html', context)
-
-# new label form
-@login_required
-def create_label(request, issue_id):
-
-    #get issue
-    observed_issue = get_object_or_404(Issue, pk=issue_id)
-    observed_project = observed_issue.project
-
-    # only project owner and collaborators can create a new label
-    if request.user != observed_project.user and not observed_project.collaborators.filter(id = request.user.id).exists():
-        return HttpResponse('Unauthorized', status=401)
-    
-    form = LabelForm(request.POST or None)
-
-    if form.is_valid():
-        label = form.save(commit=True)
-        
-        #set issue as foreign key
-        label.issue.add(observed_issue)
-        label.save()
-
-        q = LableEvent.objects.create(user=request.user, time= datetime.datetime.now(), issue=observed_issue, label=label, state="CR")
-        
-        return HttpResponseRedirect('/issue/' + str(issue_id) + '/')
-
-    context = {
-        'form' : form,
-    }
-
-    return render(request, 'uks_app/create_label.html', context)
-    
-
-# choose label
-@login_required
-def choose_label(request, issue_id):
-
-    #get issue and project
-    observed_issue = get_object_or_404(Issue, pk=issue_id)
-    observed_project = observed_issue.project
-
-    # only project owner and collaborators can choose a label
-    if request.user != observed_project.user and not observed_project.collaborators.filter(id = request.user.id).exists():
-        return HttpResponse('Unauthorized', status=401)
-    
-    project_issues = observed_project.issue_set.exclude(id=issue_id)
-
-    form = ChooseLabelForm(project_issues, observed_issue, data=request.POST or None)
-
-    if form.is_valid():
-        chosen_labels = form.cleaned_data['labels']
-        
-        for label in chosen_labels:
-            label.issue.add(observed_issue)
-            label.save()
-            q = LableEvent.objects.create(user=request.user, time= datetime.datetime.now(), issue=observed_issue, label=label, state="CR")
-        
-        return HttpResponseRedirect('/issue/' + str(issue_id) + '/')
-
-    context = {
-        'form' : form,
-        'issue' : observed_issue
-    }
-
-    return render(request, 'uks_app/choose_label.html', context)
-
-@login_required
-def choose_subissue(request, issue_id):
-
-    #get issue 
-    observed_issue = get_object_or_404(Issue, pk=issue_id)
-    observed_project = observed_issue.project
-
-    # only project owner and collaborators can choose a subissue
-    if request.user != observed_project.user and not observed_project.collaborators.filter(id = request.user.id).exists():
-        return HttpResponse('Unauthorized', status=401)
-
-    form = ChooseSubissueForm(observed_issue, data=request.POST or None)
-
-    if form.is_valid():
-    
-        chosen_issues = form.cleaned_data['issues']
-        
-        for issue in chosen_issues:
-            if issue.parent_issue == None and issue != observed_issue:
-                issue.parent_issue = observed_issue
-
-                issue.save()
-                
-                q = SubIssueEvent.objects.create(user=request.user, time= datetime.datetime.now(), issue=observed_issue, state="CR", subissue=issue)
-        
-        return HttpResponseRedirect('/issue/' + str(issue_id) + '/')
-
-    context = {
-        'form' : form,
-        'issue' : observed_issue
-    }
-
-    return render(request, 'uks_app/choose_subissue.html', context)
-
-@login_required
-def remove_subissue(request, issue_id, subissue_id):
-
-    #get issue
-    issue = get_object_or_404(Issue, id=issue_id)
-    observed_project = issue.project
-
-    # only project owner and collaborators can remove a subissue
-    if request.user != observed_project.user and not observed_project.collaborators.filter(id = request.user.id).exists():
-        return HttpResponse('Unauthorized', status=401)
-
-    #get subissue
-    subissue = get_object_or_404(Issue, id=subissue_id)
-
-    issue.subissues.remove(subissue)
-
-    q = SubIssueEvent.objects.create(user=request.user, time= datetime.datetime.now(), issue=issue, state="RE", subissue=subissue)
-
-    return HttpResponseRedirect('/issue/' + str(issue_id) + '/')
-
-# new subissue form
-@login_required
-def create_subissue(request, issue_id):
-
-    observed_issue = get_object_or_404(Issue, pk=issue_id)
-    observed_project = observed_issue.project
-
-    # only project owner and collaborators can create a subissue
-    if request.user != observed_project.user and not observed_project.collaborators.filter(id = request.user.id).exists():
-        return HttpResponse('Unauthorized', status=401)
-
-    form = IssueForm(data=request.POST or None, project=observed_project, iss=observed_issue)
-
-    if form.is_valid():
-        issue = form.save(commit=False)
-        
-        #set project as foreign key
-        issue.project = observed_project
-
-        #set parent issue
-        issue.parent_issue = observed_issue
-
-        issue.create_time = datetime.datetime.now()
-
-        issue.save()
-
-        q = SubIssueEvent.objects.create(user=request.user, time= datetime.datetime.now(), issue=observed_issue, state="CR", subissue=issue)
-        
-        return HttpResponseRedirect('/issue/' + str(issue_id) + '/')  
-    
-    context = {
-        'form' : form,
-        'cancel_url': "../"
-    }
-
-    return render(request, 'uks_app/create_update_issue.html', context)
-
-#change issue state
-def change_issue_state(request, project_id, issue_id):
-
-    #get observed project
-    observed_project = get_object_or_404(ObservedProject, id=project_id)
-
-    # only project owner and collaborators can change issue state
-    if request.user != observed_project.user and not observed_project.collaborators.filter(id = request.user.id).exists():
-        return HttpResponse('Unauthorized', status=401)
-    
-    #get issue
-    observed_issue = get_object_or_404(Issue, id=issue_id)
-
-    if observed_issue.state == "OP":
-        observed_issue.state = "CL"
-        q = IssueChange.objects.create(user=request.user, time= datetime.datetime.now(), issue=observed_issue, state="CL")
-
-    else:
-        observed_issue.state = "OP"
-        q = IssueChange.objects.create(user=request.user, time= datetime.datetime.now(), issue=observed_issue, state="OP")
-    
-    observed_issue.save()
-
-    return HttpResponseRedirect('/issue/' + str(issue_id) + '/')
-
-# assign issue
-@login_required
-def assign_issue(request, issue_id):
-
-    #get issue and project
-    observed_issue = get_object_or_404(Issue, pk=issue_id)
-    observed_project = observed_issue.project
-    
-    # only project owner and collaborators can assign a issue
-    if request.user != observed_project.user and not observed_project.collaborators.filter(id = request.user.id).exists():
-        return HttpResponse('Unauthorized', status=401)
-    
-    form = AssignIssueForm(observed_issue, data=request.POST or None)
-
-    if form.is_valid():
-        
-        chosen_user = form.cleaned_data['user'].first()
-        observed_issue.user=chosen_user
-        observed_issue.save()
-
-        assign_issue_event = AssignIssueEvent.objects.create(assigned_user=chosen_user, issue=observed_issue, time=datetime.datetime.now(), user=request.user)
-
-        return HttpResponseRedirect('/issue/' + str(issue_id) + '/')
-
-    context = {
-        'form' : form,
-        'issue' : observed_issue
-    }
-
-    return render(request, 'uks_app/assign_issue.html', context)
-
-@login_required
-def remove_label(request, label_id, issue_id):
-
-    # get issue
-    issue = get_object_or_404(Issue, id=issue_id)
-    observed_project = issue.project
-
-    if request.user != observed_project.user and not observed_project.collaborators.filter(id = request.user.id).exists():
-        return HttpResponse('Unauthorized', status=401)
-    
-    # get label
-    label = get_object_or_404(Label, id=label_id)
-
-    issue.labels.remove(label)
-
-    LableEvent.objects.create(user=request.user, time= datetime.datetime.now(), issue=issue, label=label, state="RE")
-
-    return HttpResponseRedirect('/issue/' + str(issue_id) + '/')
 
 #search projects from name
 def search_projects(request):
@@ -515,78 +199,7 @@ def hook_receiver_view(request):
     except Exception as e:
         return HttpResponse('Data is not valid', status=400)
 
-#delete project
-class ProjectDelete(DeleteView):
-    template_name = 'uks_app/delete_project.html'
-    model = ObservedProject
 
-    def dispatch(self, request, *args, **kwargs):
-       entity = get_object_or_404(ObservedProject, pk=kwargs['pk'])
-       user = request.user
-
-       if user != entity.user:
-           return HttpResponse('Unauthorized', status=401)
-       return super(ProjectDelete, self).dispatch(request, *args, **kwargs)
-
-    success_url = reverse_lazy('all_projects')
-
-#delete issue
-class IssueDelete(DeleteView):
-    template_name = 'uks_app/delete_issue.html'
-    model = Issue
-
-    def dispatch(self, request, *args, **kwargs):
-       entity = get_object_or_404(Issue, pk=kwargs['pk'])
-       user = request.user
-       observed_project = entity.project
-
-       if request.user != observed_project.user and not observed_project.collaborators.filter(id = request.user.id).exists():
-           return HttpResponse('Unauthorized', status=401)
-       return super(IssueDelete, self).dispatch(request, *args, **kwargs)
-
-    def get_success_url(self):
-        project = self.object.project
-        return reverse_lazy( 'one_project', kwargs={'pk': project.id})
-
-# all projects view
-class ProjectView(generic.ListView):
-    template_name = 'uks_app/all_projects.html'
-    context_object_name = 'project_list'
-
-    def get_queryset(self):
-        return ObservedProject.objects.filter(public=True).all()
-
-# one project detailed view
-class OneProjectView(generic.DetailView):
-    model = ObservedProject
-    template_name = 'uks_app/one_project.html'
-
-    def dispatch(self, request, *args, **kwargs):
-        entity = get_object_or_404(ObservedProject, pk=kwargs['pk'])
-        user = request.user
-
-        if not entity.public and user != entity.user and not entity.collaborators.filter(id = user.id).exists():
-            return HttpResponse('Unauthorized', status=401)
-        return super(OneProjectView, self).dispatch(request, *args, **kwargs)
-
-# one issue detailed view
-class OneIssueView(generic.DetailView):
-    model = Issue
-    template_name = 'uks_app/one_issue.html'
-
-    def get_context_data(self, **kwargs):
-        context = super(OneIssueView, self).get_context_data(**kwargs)
-        context['events'] = Event.objects.filter(issue_id=self.kwargs.get('pk')).order_by('time')
-
-        return context
-    
-    def dispatch(self, request, *args, **kwargs):
-        entity = get_object_or_404(Issue, pk=kwargs['pk'])
-        user = request.user
-
-        if not entity.project.public and user != entity.project.user and not entity.project.collaborators.filter(id = user.id).exists():
-            return HttpResponse('Unauthorized', status=401)
-        return super(OneIssueView, self).dispatch(request, *args, **kwargs)
 
 # one milestone detailed view
 class OneMilestoneView(generic.DetailView):
@@ -648,9 +261,6 @@ def create_update_milestone(request, project_id, milestone_id=None):
     }
     return render(request, 'uks_app/create_update_milestone.html', context)
     
-   
-
-
 # choose milestone
 @login_required
 def choose_milestone(request, issue_id):
